Improved ld_preload readdir

+++ tools/mapfd.cpp
@@ -8,7 +8,7 @@
 #include <mutex>
 
 using std::string;
-typedef std::map<std::string,std::pair< int, int>> Map;
+typedef std::map<std::string,std::pair< int, long>> Map;
 std::mutex fdlock;
 
 
@@ -19,7 +19,7 @@ void* map_fd_create() {
   return reinterpret_cast<void*> (new Map);
 }
 
-void map_fd_put(void* map, char* k, int v, int p) {
+void map_fd_put(void* map, char* k, int v, unsigned long p) {
   std::unique_lock<std::mutex> lck (fdlock);
   Map* m = reinterpret_cast<Map*> (map);
   std::pair<int, int> value(v, p);
@@ -28,12 +28,13 @@ void map_fd_put(void* map, char* k, int v, int p) {
   m->insert({k, value});
 }
 
-void map_fd_update_value(void* map, char* k, int v, int p) {
+void map_fd_update_value(void* map, char* k, int v, unsigned long p) {
   std::unique_lock<std::mutex> lck (fdlock);
   Map* m = reinterpret_cast<Map*> (map);
    auto search = m->find(std::string(k));
      
     if (search != m->end()) {
+        //printf("map-fd_update=%ld\n",p);
         search->second.first = v;
         search->second.second = p;
     }
@@ -45,7 +46,7 @@ void map_fd_erase(void* map, char* k) {
   m->erase(std::string(k));
 }
 
-int map_fd_search(void* map, const char* k, int *v, int *p) {
+int map_fd_search(void* map, const char* k, int *v,  unsigned long *p) {
     std::unique_lock<std::mutex> lck (fdlock);
     Map* m = reinterpret_cast<Map*> (map);
     auto search = m->find(std::string(k));
@@ -90,7 +91,7 @@ int map_fd_search_by_val_close(void* map, int v) {
     }
 
     if(remove != ""){
-        //printf("remove=%s\n",remove.c_str());
+        //printf("map_fd remove=%s fd=%d\n",remove.c_str(), v);
         m->erase(remove);
         return 1;
     }
